Add tests and cleanup
