+++ CodeMaid.UnitTests/CodeMaid.UnitTests.csproj
@@ -66,7 +66,8 @@
     <Compile Include="Helpers\SettingsMonitorTests.cs" />
     <Compile Include="MemberTypeSettingTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RoslynTests.cs" />
+    <Compile Include="Cleanup\RoslynTests.cs" />
+    <Compile Include="Cleanup\TestWorkspace.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\CodeMaid\CodeMaid.csproj">
